+ if a target file is not a zip file, then, use as it is. + if the verbose flag is not set, do not ask users intentions. Original: 347b71590aca1bd54fe47a1b23c0566f6f4e307a

+++ lalapps/src/pulsar/FDS_isolated/combiner.py
@@ -7,18 +7,18 @@ textto="%coincidences"
 
 
 ## Import standard modules to the python path
-import string, os,sys, shutil, time, os.path
+import string, os,sys, shutil, time, os.path, zipfile
 from optparse import OptionParser
 
 
 ## command line options
 parser = OptionParser()
 parser.add_option("-w", "--workdir", dest="workdir",
-                  help="Directory where the result directory will be.",
+                  help="Directory where the result file will be.",
 		  default="./")
 parser.add_option("-t", "--targetdir", dest="targetdir",
-                  help="Directory under which the data files will be.",
-                  default="./separate_targetdir/")
+                  help="Directory under which the data files are.",
+                  default="./targetdir/")
 parser.add_option("-i", "--tmpdir", dest="tmpdir",
                   help="Directory where the result files will be "
 		  "under workdir.",
@@ -26,8 +26,12 @@ parser.add_option("-i", "--tmpdir", dest="tmpdir",
 parser.add_option("-r", "--resultfile", dest="resultfile",
                   help="Name of the result file.",
 		  default="Combined")
+parser.add_option("-v", "--verbose", dest="vrbflg",
+                  help="Ask your intention for safety.",
+		  default=1)
 (options, args) = parser.parse_args()
 
+vrbflg=options.vrbflg
 
 ## set-up and normalize directory names.
 workdir=os.path.abspath(options.workdir)+"/"
@@ -39,11 +43,12 @@ if os.path.isdir(workdir) is False:
     print "Directory ",workdir ," does not exist."
 if os.path.isdir(targetdir) is False:
     print "Directory ",targetdir ," does not exist."
-## print out the directories.
-print "working directory: "+workdir
-print "target directory: "+targetdir
-print "result file name: "+resultfile
 
+## print out the directories.
+if vrbflg == 1:
+    print "working directory: "+workdir
+    print "target directory: "+targetdir
+    print "result file name: "+resultfile
 
 
 try: os.mkdir(tmpdir)
@@ -255,21 +260,23 @@ def appendfile(fromfile="",tofile=sys.stdout):
 if os.path.isfile(resultfile):		
    print "Current size of the result file."
    os.system("du -h "+resultfile)
-   uinput=raw_input("Do you want to delete [d] the existing result file or append [a] results to it?:[d/a]")
-   if uinput is "d":
-       os.remove(resultfile)
-   elif uinput is "a":
-       pass
-   else:
-       sys.exit()
+   if vrbflg == 1:
+       uinput=raw_input("Do you want to delete [d] the existing result file or append [a] results to it?:[d/a]")
+       if uinput is "d":
+           os.remove(resultfile)
+       elif uinput is "a":
+           pass
+       else:
+           sys.exit()
    
-print "size of the target directory before unzip"
+print "size of the target directory before unzip (if it is a zipped file)"
 os.system("du -h "+targetdir+" | tail -n 1")
-uinput=raw_input("Are you sure to proceed?:[y/n]")
-if uinput is "y":
-   pass
-else:
-   sys.exit()		   
+if vrbflg == 1:
+    uinput=raw_input("Are you sure to proceed?:[y/n]")
+    if uinput is "y":
+        pass
+    else:
+        sys.exit()		   
 
 
 ##---------------------------------------------------------------
@@ -282,10 +289,11 @@ for parentdir, childdirs, files in os.walk(targetdir):
 	   copiedfile=workdir+file  
 	   targetfile=os.path.join(parentdir,file)   
 	   shutil.copy2(targetfile,copiedfile) # Copy file to working dir.
-	   zipfilename=copiedfile+"_tmp.zip"   
-	   shutil.move(copiedfile,zipfilename)
-	   os.system("unzip -qa "+zipfilename)   # Unzip the copied file
-	   os.remove(zipfilename)	
+           if zipfile.is_zipfile(copiedfile) is True:
+               zipfilename=copiedfile+"_tmp.zip"   
+               shutil.move(copiedfile,zipfilename)
+               os.system("unzip -qa "+zipfilename)   # Unzip the copied file
+               os.remove(zipfilename)
 	   fileseparater(targetfile=copiedfile,textto=textto,tmpdir=tmpdir)
 ## This loop add various information to files.
 fileid=1
@@ -302,5 +310,5 @@ for parentdir, childdirs, files in os.walk(tmpdir):
 
 
 addfooterinfo(targetfile=resultfile,info="%DONE")
-
+os.system("rm -rf "+tmpdir)
